refactor: move config.?pp to Application.?pp
- Create a new folder src/Application, and then move
  src/Editor/config/config.hpp and src/Editor/config/config.cpp to it.
- Rename the new files in the Application folder to Application.hpp and
  Application.cpp respectively
