+++ src/CMakeLists.txt
@@ -18,8 +18,8 @@ target_sources(kilo
         "${PROJECT_SOURCE_DIR}/src/Editor/Editor.hpp"
         "${PROJECT_SOURCE_DIR}/src/Editor/Editor.cpp"
         
-        "${PROJECT_SOURCE_DIR}/src/Editor/config/config.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Editor/config/config.cpp"
+        "${PROJECT_SOURCE_DIR}/src/Application/Application.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Application/Application.cpp"
 
         "${PROJECT_SOURCE_DIR}/src/Offset/Offset.hpp"
