+++ tests/CMakeLists.txt
@@ -27,8 +27,8 @@ target_sources(tests
         "${PROJECT_SOURCE_DIR}/src/Editor/Editor.cpp"
         Editor.test.cpp
 
-        "${PROJECT_SOURCE_DIR}/src/Editor/config/config.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Editor/config/config.cpp"
+        "${PROJECT_SOURCE_DIR}/src/Application/Application.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Application/Application.cpp"
 
         "${PROJECT_SOURCE_DIR}/src/Offset/Offset.hpp"
